1. drop meta cache 2. force span.fx_area_marker to be invisible 3. fixed wysiwyg image uploader script

+++ Admin/Controller/Field.php
@@ -283,6 +283,9 @@ class Field extends Admin
         foreach ($ids as $id) {
             try {
                 $field = fx::data($es, $id);
+                if (!$field) {
+                    continue;
+                }
                 $field->delete();
                 if ($field['component_id']) {
                     $com = fx::data('component', $field['component_id']);
